+++ pkg/models/deleted-spot-info-model.go
@@ -94,7 +94,7 @@ func (cfg *DeletedSpotInfoConfiguration) InitJob() func() {
 			Id string `json:"ID"`
 		}
 		var budgets []Budget
-		months := map[int][]time.Time{}
+		months := map[int64][]time.Time{}
 		badgerBudgets := storage.Open(DbBudgets)
 		err = badgerBudgets.Find(&budgets, badgerhold.Where("Month").Ge(-1))
 		if err != nil {
